Move AdditionalAdditions to FoundationAdditions.
