+++ FoundationAdditions/AdditionalAdditions.swift
@@ -343,20 +343,3 @@ public extension UnsafeMutableBufferPointer {
 		self = UnsafeMutableBufferPointer(start: start, count: start.distance(to: toIterate))
 	}
 }
-
-/// For `OSType` values that can be represented as a four-character string with values in the
-/// *Mac OS Roman* string encoding.
-public protocol OSTypeConvertable: RawRepresentable where RawValue == OSType {
-	/// The value's string representation.
-	///
-	/// The value must be representable in the Mac OS Roman encoding.
-	/// If the value contains any escape characters (*0x00*..*0x20*)
-	/// or is unable to be seen as a Mac OS Roman string, will be an empty string.
-	var stringValue: String { get }
-}
-
-public extension OSTypeConvertable {
-	var stringValue: String {
-		return OSTypeToString(self.rawValue) ?? ""
-	}
-}
