+++ SwiftAdditions/MacTypesAdditions.swift
@@ -14,6 +14,7 @@ import AppKit.NSWorkspace
 import AppKit.NSImage
 #endif
 import CoreGraphics
+import FoundationAdditions
 
 
 /// Converts an `OSType` to a `String` value. May return `nil`.
@@ -1000,6 +1001,23 @@ public enum CarbonGenericFinderIcons: OSType, OSTypeConvertable, OSTypeIconConve
 }
 
 #endif
+
+/// For `OSType` values that can be represented as a four-character string with values in the
+/// *Mac OS Roman* string encoding.
+public protocol OSTypeConvertable: RawRepresentable where RawValue == OSType {
+	/// The value's string representation.
+	///
+	/// The value must be representable in the Mac OS Roman encoding.
+	/// If the value contains any escape characters (*0x00*..*0x20*)
+	/// or is unable to be seen as a Mac OS Roman string, will be an empty string.
+	var stringValue: String { get }
+}
+
+public extension OSTypeConvertable {
+	var stringValue: String {
+		return OSTypeToString(self.rawValue) ?? ""
+	}
+}
 
 /// Old QuickDraw text styles.
 public struct QuickDrawStyle : OptionSet {
